refactor(product): remove duplicated admin legacy files
